ai_service/requirements.txt
fastapi
uvicorn
openai
supabase
langchain
pydantic
python-dotenv
tiktoken
pypdf
sqlalchemy
mysql-connector-python
python-multipart
jose
passlib[bcrypt]
